+++ x/finality/keeper/msg_server.go
@@ -238,13 +238,8 @@ func (ms msgServer) ShouldAcceptSigForHeight(ctx context.Context, block *types.I
 func (ms msgServer) CommitPubRandList(goCtx context.Context, req *types.MsgCommitPubRandList) (*types.MsgCommitPubRandListResponse, error) {
 	defer telemetry.ModuleMeasureSince(types.ModuleName, time.Now(), types.MetricsKeyCommitPubRandList)
 
-	// To avoid public randomness reset,
-	// check for overflow when doing (StartHeight + NumPubRand)
-	if req.StartHeight >= (req.StartHeight + req.NumPubRand) {
-		return nil, types.ErrOverflowInBlockHeight.Wrapf(
-			"public rand commit start block height: %d is equal or higher than (start height + num pub rand) %d",
-			req.StartHeight, req.StartHeight+req.NumPubRand,
-		)
+	if err := req.ValidateBasic(); err != nil {
+		return nil, err
 	}
 
 	ctx := sdk.UnwrapSDKContext(goCtx)